update dashbaord view
