--- admin/templates/sales_analytics/debug_test.html
@@ -1,7 +1,0 @@
-{% extends 'base_analytics.html' %}
-{% load static %}
-
-{% block title_analytics %}
-    <h1>Debug and Test</h1>
-{% endblock title_analytics %}
-
