--- admin/templates/sales_analytics/products_usage.html
@@ -1,6 +1,0 @@
-{% extends 'base_analytics.html' %}
-{% load static %}
-
-{% block title_analytics %}
-    <h1>OSF Products Usage</h1>
-{% endblock title_analytics %}
