--- admin/templates/sales_analytics/products_view.html
@@ -1,17 +1,0 @@
-{% extends 'base_analytics.html' %}
-{% load static %}
-
-{% block title_analytics %}
-    <h1>OSF Products View</h1>
-{% endblock title_analytics %}
-
-{% block content_analytics %}
-    <div id='keen-chart-osf-products-view-mt2'>...</div>
-    <div id='keen-chart-osf-products-view-mee'>...</div>
-    <div id='keen-chart-osf-products-view-pre'>...</div>
-    <div id='keen-chart-osf-products-view-ins'>...</div>
-{% endblock content_analytics %}
-
-{% block bottom_js %}
-    <script src="{% static 'public/js/sa-products-view.js' %}"></script>
-{% endblock bottom_js %}
